Added:User id now stored in notes to have relationship between them

+++ node-easy-notes-app/app/models/note.model.js
@@ -3,7 +3,7 @@ const mongoose = require('mongoose');
 const NoteSchema = mongoose.Schema({
     title: String,
     content: String,
-    //userId: { type: mongoose.Schema.Types.ObjectId, ref: "User" },
+    userId: { type: mongoose.Schema.Types.ObjectId, ref: "User" }
 }, {
     timestamps: true
 });
@@ -12,55 +12,84 @@ const myNote = mongoose.model("Note", NoteSchema);
 
 class noteModel {
     //creates a note and saves it in database
-    createNote = (title, content, callback) => {
+    createNote = (title, content, userId, callback) => {
       const note = new myNote({
         title: title,
         content: content,
-        //userId: userId,
+        userId: userId,
       });
-      // Save Note in the database
       return note.save((err, data) => {
         return err ? callback(err, null) : callback(null, data);
       });
     };
   
     // Retrieve and return all notes from the database.
-    findAll = (callback) => {
-      return myNote.find((err, data) => {
-        return err ? callback(err, null) : callback(null, data);
+    findAll = (userId,callback) => {
+      return myNote
+      .find({ userId: userId })
+      .populate({
+        path: "userId",
+        select: ["name", "phoneNumber", "email"],
+      })
+      .exec((error, data) => {
+        return error ? callback(error, null) : callback(null, data);
       });
     };
   
     // Find a single note with a noteId
-    findOne = (noteId, callback) => {
-      myNote.findById(noteId, (err, data) => {
-        return err ? callback(err, null) : callback(null, data);
+    findOne = (userId,noteId, callback) => {
+      return myNote.findOne({ userId: userId, _id: noteId }, (error, data) => {
+        if (error) {
+          return callback(error, null);
+        }
+        if (!data) {
+          return callback("You dont have access to this note", null);
+        } else {
+          return callback(null, data);
+        }
       });
     };
     
   
     // Update a note identified by the noteId in the request
-    updateNote = (noteId, title, content, callback) => {
-      // Find note and update it with the request body
-      myNote.findByIdAndUpdate(
-        noteId,
+    updateNote = (userId, noteId, body, callback) => {
+      return myNote.findOneAndUpdate(
+        { 
+          userId: userId, 
+          _id: noteId 
+        },
         {
-          title: title || "Untitled Note",
-          content: content,
+          title: body.title,
+          content: body.content,
         },
         { new: true },
-        (err, data) => {
-          return err ? callback(err, null) : callback(null, data);
+        (error, data) => {
+          if (error) {
+            return callback(error, null);
+          }
+          if (!data) {
+            return callback("You dont have access to this note", null);
+          } else {
+            return callback(null, data);
+          }
         }
       );
     };
   
     // Delete a note with the specified noteId in the request
-    deleteOne = (noteId, callback) => {
-      myNote.findByIdAndRemove(noteId, (err, data) => {
-        return err ? callback(err, null) : callback(null, data);
+    deleteOne = (userId,noteId, callback) => {
+      myNote.findOneAndRemove({ userId: userId, _id: noteId }, (error, data) => {
+        if (error) {
+          return callback(error, null);
+        }
+        if (!data) {
+          return callback("You dont have access to this note", null);
+        } else {
+          return callback(null, data);
+        }
       });
     };
-  }
+
+}
   
 module.exports = new noteModel();
